Button & Badge file rename
